Qol fixes (#195)
## Description
Escape now pauses when in game. Character select and ready up uses the
bound up down and punch keys. Menus are fps capped to monitor refresh
rate by enabling vsync on state transitions

## Issue(s) Completed by this PR
#177 #192 #180 

## Changes Made
```
Outline the specific changes made in this pull request. (if there are too many to list you should've made multiple PRs)
```

## Screenshots (if applicable)

## Testing
Manually tested on linux and windows

### Bugs


## Checklist 
```
Both the requestee and the reviewer should review this checklist
```
- [X] Code Compiles on your machine
- [X] PR Filled in (remove the code formatted blocks)
- [X] Code is appropriately compartmentalized with 
- [X] feature has no/minimal interference for other functions
- [X] Code is well documented with comments
- [X] Code is readable
- [X] Tests (if included) pass

## Additional Notes
The pointer in the Character select menu is still R and X as they are
textures

---------

Co-authored-by: nicholasung <nicholas.ung0@gmail.com>
Co-authored-by: Armaan <armaan070903@gmail.com>